backend/submission-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
aiofiles
pymysql
python-multipart
pydantic
pydantic[email]
pydantic-settings
python-dotenv
httpx
requests
cryptography
pypdf
redis>=5.0.0
prometheus-fastapi-instrumentator==7.0.0
pika==1.3.2
kafka-python==2.0.2